Split `KeyLogFile` into its own submodule & make it optional.
Split `KeyLogFile` into its own module to make it clearer that
`KeyLog` doesn't depend on `std::{env,fs,io,path,sync}`.

Add a feature flag to allow disabling the `KeyLogFile` feature. Make it
a default feature for API backward compatibility. In the next breaking
release, we should consider making it a non-default feature.

Use `git diff main:rustls/src/keylog.rs rustls/src/keylogfile.rs` to
verify that none of the `KeyLogFile` code has changed at all.
